Divide tests into two logical parts (loading - backend functions), add test for RepairTile
